Rename DiscoveryBeta -> Discovery
