+++ src/Discovery/DiscoveryConfig.d.ts
