+++ src/Discovery/__mocks__/mock_config.json
@@ -1,5 +1,5 @@
 {
-    "pageTitle": "My Special Test Discovery Page",
+    "page_title": "My Special Test Discovery Page",
     "features": {
       "exploration_integration": {
         "enabled": false
